+++ docs/conf.py
@@ -12,7 +12,7 @@ copyright = "2026, Steadlith contributors"
 try:
     release = version("steadlith")
 except PackageNotFoundError:
-    release = "1.0.0"
+    release = "1.0.1"
 version = ".".join(release.split(".")[:2])
 
 extensions = [